ManualSymbols - Clean up code - Ignore Unitprice of 0
