+++ GhostfolioSidekick/Ghostfolio/API/GhostfolioAPI.cs
@@ -26,15 +26,9 @@ namespace GhostfolioSidekick.Ghostfolio.API
 			IMemoryCache memoryCache,
 			ILogger<GhostfolioAPI> logger)
 		{
-			if (settings is null)
-			{
-				throw new ArgumentNullException(nameof(settings));
-			}
+			ArgumentNullException.ThrowIfNull(settings);
 
-			if (memoryCache is null)
-			{
-				throw new ArgumentNullException(nameof(memoryCache));
-			}
+			ArgumentNullException.ThrowIfNull(memoryCache);
 
 			this.settings = settings;
 			this.memoryCache = memoryCache;
@@ -580,7 +574,7 @@ namespace GhostfolioSidekick.Ghostfolio.API
 				logger.LogDebug($"Skipping empty transaction {activity.Date} {activity.SymbolProfile.Symbol} {activity.Quantity} {activity.Type}");
 			}
 
-			if (activity.Type == Contract.ActivityType.IGNORE)
+			if (activity.Type == Ghostfolio.Contract.ActivityType.IGNORE)
 			{
 				logger.LogDebug($"Skipping ignore transaction {activity.Date} {activity.SymbolProfile.Symbol} {activity.Quantity} {activity.Type}");
 			}
@@ -652,13 +646,9 @@ namespace GhostfolioSidekick.Ghostfolio.API
 			o["fee"] = activity.Fee;
 			o["quantity"] = activity.Quantity;
 
-			if (activity.Type == Contract.ActivityType.INTEREST)
+			if (activity.Type == Ghostfolio.Contract.ActivityType.INTEREST)
 			{
 				o["symbol"] = "Interest";
-			}
-			else if (activity.Type == Contract.ActivityType.FEE)
-			{
-				o["symbol"] = "Fee";
 			}
 			else
 			{
@@ -716,8 +706,7 @@ namespace GhostfolioSidekick.Ghostfolio.API
 		private bool AreEquals(Contract.Activity fo, Contract.Activity eo)
 		{
 			return
-				(fo.SymbolProfile?.Symbol == eo.SymbolProfile?.Symbol ||
-					fo.Type == Contract.ActivityType.INTEREST || fo.Type == Contract.ActivityType.FEE) && // Interest & Fee create manual symbols
+				(fo.SymbolProfile?.Symbol == eo.SymbolProfile?.Symbol || fo.Type == Ghostfolio.Contract.ActivityType.INTEREST) && // Interest create manual symbols
 				fo.Quantity == eo.Quantity &&
 				fo.UnitPrice == eo.UnitPrice &&
 				fo.Fee == eo.Fee &&
