+++ GhostfolioSidekick/MarketDataMaintainer/CreateManualSymbolTask.cs
@@ -1,7 +1,6 @@
-﻿using GhostfolioSidekick.Configuration;
-using GhostfolioSidekick.FileImporter;
+﻿using GhostfolioSidekick.FileImporter;
 using GhostfolioSidekick.Ghostfolio.API;
-using GhostfolioSidekick.Model;
+using GhostfolioSidekick.MarketDataMaintainer.Actions;
 using Microsoft.Extensions.Logging;
 
 namespace GhostfolioSidekick.MarketDataMaintainer
@@ -10,8 +9,7 @@ namespace GhostfolioSidekick.MarketDataMaintainer
 	{
 		private readonly ILogger<FileImporterTask> logger;
 		private readonly IGhostfolioAPI api;
-		private readonly ConfigurationInstance configurationInstance;
-		private int counter = -1;
+		private readonly CreateManualSymbol action;
 
 		public int Priority => 2;
 
@@ -20,14 +18,12 @@ namespace GhostfolioSidekick.MarketDataMaintainer
 			IGhostfolioAPI api,
 			IApplicationSettings applicationSettings)
 		{
-			if (applicationSettings is null)
-			{
-				throw new ArgumentNullException(nameof(applicationSettings));
-			}
+			ArgumentNullException.ThrowIfNull(applicationSettings);
 
 			this.logger = logger ?? throw new ArgumentNullException(nameof(logger));
 			this.api = api ?? throw new ArgumentNullException(nameof(api));
-			this.configurationInstance = applicationSettings.ConfigurationInstance;
+
+			action = new CreateManualSymbol(api, applicationSettings.ConfigurationInstance);
 		}
 
 		public async Task DoWork()
@@ -36,7 +32,7 @@ namespace GhostfolioSidekick.MarketDataMaintainer
 
 			try
 			{
-				await ManageManualSymbols();
+				await action.ManageManualSymbols();
 			}
 			catch (NotAuthorizedException)
 			{
@@ -45,106 +41,6 @@ namespace GhostfolioSidekick.MarketDataMaintainer
 			}
 
 			logger.LogInformation($"{nameof(CreateManualSymbolTask)} Done");
-		}
-
-		private async Task ManageManualSymbols()
-		{
-			var marketData = (await api.GetMarketData()).ToList();
-			var activities = (await api.GetAllActivities()).ToList();
-
-			var symbolConfigurations = configurationInstance.Symbols;
-			foreach (var symbolConfiguration in symbolConfigurations)
-			{
-				var manualSymbolConfiguration = symbolConfiguration.ManualSymbolConfiguration;
-				if (manualSymbolConfiguration == null)
-				{
-					continue;
-				}
-
-				await AddOrUpdateSymbol(symbolConfiguration, manualSymbolConfiguration);
-				await SetKnownPrices(symbolConfiguration, marketData, activities);
-			}
-		}
-
-		private async Task SetKnownPrices(SymbolConfiguration symbolConfiguration, List<MarketDataList> marketData, List<Activity> activities)
-		{
-			var mdi = marketData.SingleOrDefault(x => x.AssetProfile.Symbol == symbolConfiguration.Symbol);
-			if (mdi == null || mdi.AssetProfile.ActivitiesCount <= 0)
-			{
-				return;
-			}
-
-			var md = await api.GetMarketData(mdi.AssetProfile.Symbol, mdi.AssetProfile.DataSource);
-			var activitiesForSymbol = activities.Where(x => x.Asset?.Symbol == mdi.AssetProfile.Symbol && IsBuyOrSell(x.ActivityType)).ToList();
-
-			if (!activitiesForSymbol.Any())
-			{
-				return;
-			}
-
-			var sortedActivities = activitiesForSymbol.OrderBy(x => x.Date).ToList();
-
-			for (var i = 0; i < sortedActivities.Count(); i++)
-			{
-				var fromActivity = sortedActivities[i];
-				Activity toActivity;
-
-				if (i + 1 < sortedActivities.Count())
-				{
-					toActivity = sortedActivities[i + 1];
-				}
-				else
-				{
-					toActivity = new Activity { Date = DateTime.Today.AddDays(1), UnitPrice = fromActivity.UnitPrice };
-				}
-
-				for (var date = fromActivity.Date; date <= toActivity.Date; date = date.AddDays(1))
-				{
-					var a = (decimal)(date - fromActivity.Date).TotalDays;
-					var b = (decimal)(toActivity.Date - date).TotalDays;
-					var percentage = a / (a + b);
-					decimal amountFrom = fromActivity.UnitPrice.Amount;
-					decimal amountTo = toActivity.UnitPrice.Amount;
-					var expectedPrice = amountFrom + (percentage * (amountTo - amountFrom));
-
-					var price = md.MarketData.FirstOrDefault(x => x.Date.Date == date.Date);
-
-					var diff = (price?.MarketPrice ?? 0) - expectedPrice;
-					if (Math.Abs(diff) >= 0.0000001M)
-					{
-						await api.SetMarketPrice(md.AssetProfile, new Money(fromActivity.UnitPrice.Currency, expectedPrice, date));
-					}
-				}
-			}
-		}
-
-		private async Task AddOrUpdateSymbol(SymbolConfiguration symbolConfiguration, ManualSymbolConfiguration? manualSymbolConfiguration)
-		{
-			var symbol = await api.FindSymbolByIdentifier(
-				symbolConfiguration.Symbol,
-				null,
-				new AssetClass?[] { Utilities.ParseEnum<AssetClass>(manualSymbolConfiguration.AssetClass) },
-				new AssetSubClass?[] { Utilities.ParseEnum<AssetSubClass>(manualSymbolConfiguration.AssetSubClass) });
-			if (symbol == null)
-			{
-				await api.CreateManualSymbol(new SymbolProfile
-				{
-					AssetClass = Utilities.ParseEnum<AssetClass>(manualSymbolConfiguration.AssetClass),
-					AssetSubClass = Utilities.ParseEnum<AssetSubClass>(manualSymbolConfiguration.AssetSubClass),
-					Currency = CurrencyHelper.ParseCurrency(manualSymbolConfiguration.Currency),
-					DataSource = "MANUAL",
-					Name = manualSymbolConfiguration.Name,
-					Symbol = symbolConfiguration.Symbol,
-					ISIN = manualSymbolConfiguration.ISIN
-				});
-			}
-
-			// TODO: update on difference???
-		}
-
-		private bool IsBuyOrSell(ActivityType activityType)
-		{
-			return activityType == ActivityType.Buy || activityType == ActivityType.Sell;
 		}
 	}
 }
